# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KeyDrive API",
        "contact": {
            "name": "KeyDrive Team",
            "url": "https://github.com/keydrive/keydrive/issues",
            "email": "thomas.biesaart@protonmail.com"
        },
        "version": "development"
    },
    "paths": {
...
        "/api/libraries/{libraryId}/entries": {
            "get": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Search the collection of files and folders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The parent folder",
                        "name": "parent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The entry path. If this value is set, all other parameters are ignored and a maximum of 1 value is returned.",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The library id",
                        "name": "libraryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.FileInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Create a new file or folder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The library id",
                        "name": "libraryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The name of the new entry. Required when creating a folder.",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "The path to the parent folder. When missing this creates a file or folder in the root of the library.",
                        "name": "parent",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "The file contents. Required when creating a file.",
                        "name": "data",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FileInfo"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Delete a file or folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The url encoded path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The library id",
                        "name": "libraryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/api/libraries/{libraryId}/entries/download": {
            "post": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Create a download token",
                "parameters": [
                    {
                        "description": "The file to create a download token for",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateDownloadTokenDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.DownloadTokenDTO"
                        }
                    }
                }
            }
        },
...
        "/api/system/browse": {
            "post": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Browse the system storage to find paths",
                "parameters": [
                    {
                        "description": "The request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.BrowseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.BrowseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user": {
            "get": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get the currently authenticated user information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Update the currently authenticated user",
                "parameters": [
                    {
                        "description": "The changes",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.BrowseRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "controller.BrowseResponse": {
            "type": "object",
            "properties": {
                "folders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.BrowseResponseFolder"
                    }
                },
                "parent": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "controller.BrowseResponseFolder": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "controller.CreateDownloadTokenDTO": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
...
        "controller.DownloadTokenDTO": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "controller.UpdateUserDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "service.FileInfo": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth2/token"
        }
    }
}